reverted changes

+++ sobel_cpu.cpp
@@ -48,20 +48,20 @@ sobel_filtered_pixel(float *s, int i, int j , int ncols, int nrows, float *gx, f
 
    // ADD CODE HERE: add your code here for computing the sobel stencil computation at location (i,j)
    // of input s, returning a float
-   float calculatedGx, calculatedGy = 0.0;
+   int calculatedGx, calculatedGy = 0;
    for(int x = 0; x < 3; x++)
    {
       for(int y = 0; y < 3; y++)
       {
-          if(i == 0 || i == nrows - 1 || j == 0 || j == ncols - 1)
+         // if we reached an edge
+         if(i == 0 || i == nrows - 1 || j == 0 || j == ncols - 1)
             return 0;  
-         
          calculatedGx += s[(i + x - 1) * ncols + (j + y - 1)] * gx[x * 3 + y];
          calculatedGy += s[(i + x - 1) * ncols + (j + y - 1)] * gy[x * 3 + y];
       }
    }
    t = sqrt(calculatedGx * calculatedGx + calculatedGy * calculatedGy);
-
+   
    return t;
 }
